cambio a clase abstracta
